+++ CHANGELOG.md
@@ -19,6 +19,8 @@ NEXT RELEASE
   - Handle some really bad WeChat useragents.
   - Handle the Despegar App
   - Detect Opera GX
+  - Detect Fuchsia OS
+  - Better Brand extraction
 - Updated UDF dependencies
   - Trino 422 (it needs 421 or newer!)